Fix Slythe weaponry.
In DSotSC-oBG1 Slythe wields a Shadow Crossbow and a Scorcher cutlass.
The Scorcher is still in the mod and thereby assigned to him, while the
Shadow Crossbow is substituted back to the Light Crossbow +1 it's based
on. (The Shadow Crossbow has different weight, weapon speed and
item usability flags only.)