Only attempt to edit hidespl.2da if it exists.
On classic, hidespl.2da was introduced by ToBEx. While SR does install
ToBEx and this file as part of itself, there might be other mods besides
it attempting to introduce these two spells on that platform.

This was something I accidentally introduced, sorry. SR including ToBEx
was the reason my compat test didn't triggered an error
on that combination on classic back in the previous PR.